+++ tests/c/chfl_trajectory_read.c
@@ -22,15 +22,16 @@ int main(){
     assert(!chfl_frame_atoms_count(frame, &natoms));
     assert(natoms == 297);
 
-    float pos_0[3] = {0.417219, 8.303366, 11.737172};
-    float pos_124[3] = {5.099554, -0.045104, 14.153846};
-    float pos[297][3];
+    float positions_0[3] = {0.417219, 8.303366, 11.737172};
+    float positions_124[3] = {5.099554, -0.045104, 14.153846};
+    float (*positions)[3] = NULL;
 
     // Check positions in the first frame
-    assert(!chfl_frame_positions(frame, pos, 297));
+    assert(!chfl_frame_positions(frame, &positions, &natoms));
+    assert(natoms == 297);
     for (unsigned i=0; i<3; i++){
-        assert(pos[0][i] == pos_0[i]);
-        assert(pos[124][i] == pos_124[i]);
+        assert(positions[0][i] == positions_0[i]);
+        assert(positions[124][i] == positions_124[i]);
     }
 
     // Check topology in the first frame
@@ -65,13 +66,14 @@ int main(){
     assert(c == 30.0);
     assert(!chfl_cell_free(cell));
 
-    pos_0[0] = 0.761277;  pos_0[1] = 8.106125;   pos_0[2] = 10.622949;
-    pos_124[0] = 5.13242; pos_124[1] = 0.079862; pos_124[2] = 14.194161;
+    positions_0[0] = 0.761277;  positions_0[1] = 8.106125;   positions_0[2] = 10.622949;
+    positions_124[0] = 5.13242; positions_124[1] = 0.079862; positions_124[2] = 14.194161;
 
-    assert(!chfl_frame_positions(frame, pos, 297));
+    assert(!chfl_frame_positions(frame, &positions, &natoms));
+    assert(natoms == 297);
     for (unsigned i=0; i<3; i++){
-        assert(pos[0][i] == pos_0[i]);
-        assert(pos[124][i] == pos_124[i]);
+        assert(positions[0][i] == positions_0[i]);
+        assert(positions[124][i] == positions_124[i]);
     }